requirements.txt
Django==5.0.4
django-cors-headers==4.3.1
django-jazzmin==3.0.1
django-tinymce==4.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
python-decouple==3.8
python-dotenv==1.0.1